+++ src/client/components/Dropzone/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Dropzone'
