+++ src/client/pages/OriginalDataPoint/components/DataSources/References/AddFromRepository/index.ts
@@ -1,0 +1,1 @@
+export { default } from './AddFromRepository'
