+++ src/client/pages/OriginalDataPoint/components/DataSources/References/References.tsx
@@ -4,18 +4,14 @@ import { useTranslation } from 'react-i18next'
 import type { Jodit } from 'jodit/types/jodit'
 import { Objects } from 'utils/objects'
 
-import { ApiEndPoint } from 'meta/api/endpoint'
-import { AreaCode } from 'meta/area'
-import { AssessmentFile, AssessmentName, CycleName, OriginalDataPoint } from 'meta/assessment'
+import { OriginalDataPoint } from 'meta/assessment'
 import { Topics } from 'meta/messageCenter'
 
-import { useAssessmentFiles, useGetAssessmentFiles } from 'client/store/ui/assessmentFiles/hooks'
-import { useCountryRouteParams } from 'client/hooks/useRouteParams'
-import ButtonCheckBox from 'client/components/ButtonCheckBox'
 import EditorWYSIWYG from 'client/components/EditorWYSIWYG'
 import MarkdownPreview from 'client/components/MarkdownPreview'
-import { Modal, ModalBody, ModalFooter, ModalHeader } from 'client/components/Modal'
 import ReviewIndicator from 'client/components/ReviewIndicator'
+
+import AddFromRepository from './AddFromRepository'
 
 type ButtonType = Jodit['options']['buttons'][0]
 
@@ -24,102 +20,6 @@ type ReferencesProps = {
   updateOriginalDataPoint: (originalDataPoint: OriginalDataPoint) => void
   disabled: boolean
   reviewIndicator: boolean
-}
-
-// TODO: Move to meta/assessment(cycle?)/files
-// TODO: Cleanup
-const getLink = (url: string, text: string) => <a href={`"${url}"`}>{text}</a>
-const getLinkText = (url: string, text: string) => `<a href="${url}">${text}</a>`
-const getAssessmentFileLink = (
-  assessmentFile: AssessmentFile,
-  assessmentName: AssessmentName,
-  cycleName: CycleName,
-  countryIso: AreaCode,
-  asText = false
-) => {
-  const text = assessmentFile.fileName
-  const url = `${ApiEndPoint.File.Assessment.one(
-    assessmentFile.uuid
-  )}?assessmentName=${assessmentName}&cycleName=${cycleName}&countryIso=${countryIso}`
-
-  if (asText) return getLinkText(url, text)
-  return getLink(url, text)
-}
-
-type AddFromRepositoryProps = { isOpen: boolean; onClose: (x: string) => void }
-
-// TODO: Move to AddFromRepository/AddFromRepository.tsx
-const AddFromRepository: React.FC<AddFromRepositoryProps> = (props: AddFromRepositoryProps) => {
-  const { isOpen, onClose } = props
-  const { assessmentName, cycleName, countryIso } = useCountryRouteParams()
-  const [selectedFiles, setSelectedFiles] = useState([])
-
-  useGetAssessmentFiles()
-  const assessmentFiles = useAssessmentFiles()
-
-  if (assessmentFiles.length) return null
-
-  const countryFiles = assessmentFiles[countryIso] || []
-
-  const allSelected = selectedFiles.length === countryFiles.length
-  const isChecked = (uuid: string) => selectedFiles.some((selectedFile) => selectedFile.uuid === uuid)
-
-  const onClick = (uuid: string) => {
-    if (isChecked(uuid)) setSelectedFiles(selectedFiles.filter((selectedFile) => selectedFile.uuid !== uuid))
-    else setSelectedFiles([...selectedFiles, countryFiles.find((file) => file.uuid === uuid)])
-  }
-
-  const onClickAll = () => {
-    if (selectedFiles.length === countryFiles.length) setSelectedFiles([])
-    else setSelectedFiles(countryFiles)
-  }
-
-  return (
-    <Modal isOpen={isOpen}>
-      <ModalHeader>
-        {/*
-        TODO: Change or translate
-        */}
-        <h1>Select files</h1>
-      </ModalHeader>
-      <ModalBody>
-        {/* TODO: label: Change or translate */}
-        <ButtonCheckBox onClick={onClickAll} checked={allSelected} label="All" suffix="" />
-        <div className="divider" />
-
-        <div className="export__form-section-variables">
-          {countryFiles.map((assessmentFile) => {
-            return (
-              <ButtonCheckBox
-                key={assessmentFile.uuid}
-                checked={isChecked(assessmentFile.uuid)}
-                label={getAssessmentFileLink(assessmentFile, assessmentName, cycleName, countryIso)}
-                onClick={() => onClick(assessmentFile.uuid)}
-              />
-            )
-          })}
-        </div>
-      </ModalBody>
-      <ModalFooter>
-        <button
-          type="button"
-          className="btn btn--primary"
-          onClick={() =>
-            onClose(
-              selectedFiles
-                .map((file) => getAssessmentFileLink(file, assessmentName, cycleName, countryIso, true))
-                .join(' ')
-            )
-          }
-        >
-          {/*
-          TODO: Change or translate
-          */}
-          Close
-        </button>
-      </ModalFooter>
-    </Modal>
-  )
 }
 
 const References: React.FC<ReferencesProps> = (props) => {
