+++ src/client/store/ui/assessmentFiles/actions/upload.ts
@@ -3,12 +3,12 @@ import axios from 'axios'
 
 import { ApiEndPoint } from 'meta/api/endpoint'
 import { CycleParams } from 'meta/api/request'
-import { CountryIso } from 'meta/area'
+import { AreaCode } from 'meta/area'
 import { AssessmentFile } from 'meta/assessment'
 
 type Params = CycleParams & {
   file: File | null
-  fileCountryIso?: CountryIso
+  fileCountryIso?: AreaCode
 }
 
 export const upload = createAsyncThunk<AssessmentFile, Params>('assessmentFiles/put/upload', async (params) => {
