+++ src/client/store/ui/assessmentFiles/hooks/index.ts
@@ -1,7 +1,7 @@
 import { useCallback, useEffect } from 'react'
 import { useTranslation } from 'react-i18next'
 
-import { CountryIso } from 'meta/area'
+import { AreaCode, CountryIso } from 'meta/area'
 
 import { useAppDispatch, useAppSelector } from 'client/store'
 import { AssessmentFilesActions } from 'client/store/ui/assessmentFiles/slice'
@@ -37,7 +37,7 @@ export const useUpdateAssessmentFiles = () => {
   const { t } = useTranslation()
 
   return useCallback(
-    (props: { fileCountryIso?: CountryIso; file: File }) => {
+    (props: { fileCountryIso?: CountryIso | AreaCode; file: File }) => {
       const { fileCountryIso, file } = props
       dispatch(
         AssessmentFilesActions.upload({
